Change variable type

+++ hamming/src/hamming.c
@@ -1,18 +1,18 @@
 #include "hamming.h"
-#include "string.h"
+#include <string.h>
 
-int compute(const char *lhs, const char *rhs)
+long long compute(const char *lhs, const char *rhs)
 {
-  int l_len = strlen(lhs);
-  int r_len = strlen(rhs);
+  long long l_len = strlen(lhs);
+  long long r_len = strlen(rhs);
 
   if (l_len != r_len)
   {
     return -1;
   }
 
-  int counter = 0;
-  for (int i = 0 ; i < l_len ; i++)
+  long long counter = 0;
+  for (long long i = 0 ; i < l_len ; i++)
   {
     if (lhs[i] != rhs[i])
     {
@@ -21,3 +21,6 @@ int compute(const char *lhs, const char *rhs)
   }
   return counter;
 }
+
+// Followed a mentor's suggestion to change int for long long to
+// represent the length of long strings (up to 9,223,372,036,854,775,807)
